misc: Add nix-shell boilerplate
This enables users of nix-shell to have most of required development
environment by changing directory into a checkout of this repo. Or by
explicitly running nix-shell.

While this is not fully complete (e.g. no gopls installed), and expect
tools like docker to be pre-installed this can be still be useful.

Signed-off-by: Dhi Aurrahman <dio@rockybars.com>